+++ liquid/builtin/tags/if_tag.py
@@ -2,6 +2,8 @@
 from __future__ import annotations
 import sys
 
+from io import StringIO
+
 from typing import Optional
 from typing import List
 from typing import TextIO
@@ -51,6 +53,7 @@ class IfNode(Node):
         "consequence",
         "conditional_alternatives",
         "alternative",
+        "forced_output",
     )
 
     def __init__(
@@ -66,6 +69,16 @@ class IfNode(Node):
         self.consequence = consequence
         self.conditional_alternatives = conditional_alternatives
         self.alternative = alternative
+
+        self.forced_output = any(
+            getattr(n, "forced_output", False)
+            for n in (
+                self.consequence,
+                *self.conditional_alternatives,
+                self.alternative,
+            )
+            if n
+        )
 
     def __str__(self) -> str:
         buf = [
@@ -83,34 +96,48 @@ class IfNode(Node):
         return f"IfNode(tok={self.tok}, condition='{self.condition}')"
 
     def render_to_output(self, context: Context, buffer: TextIO) -> Optional[bool]:
+        # This intermediate buffer is used to detect and possibly suppress blocks that,
+        # when rendered, contain only whitespace.
+        buf = StringIO()
+
         if self.condition.evaluate(context):
-            rendered = self.consequence.render(context, buffer)
+            rendered = self.consequence.render(context, buf)
         else:
             rendered = False
             for alt in self.conditional_alternatives:
-                if alt.render(context, buffer):
+                if alt.render(context, buf):
                     rendered = True
                     break
 
             if not rendered and self.alternative:
-                return self.alternative.render(context, buffer)
+                rendered = self.alternative.render(context, buf)
+
+        val = buf.getvalue()
+        if self.forced_output or not val.isspace():
+            buffer.write(val)
 
         return rendered
 
     async def render_to_output_async(
         self, context: Context, buffer: TextIO
     ) -> Optional[bool]:
+        buf = StringIO()
+
         if await self.condition.evaluate_async(context):
-            rendered = await self.consequence.render_async(context, buffer)
+            rendered = await self.consequence.render_async(context, buf)
         else:
             rendered = False
             for alt in self.conditional_alternatives:
-                if await alt.render_async(context, buffer):
+                if await alt.render_async(context, buf):
                     rendered = True
                     break
 
             if not rendered and self.alternative:
-                return await self.alternative.render_async(context, buffer)
+                rendered = await self.alternative.render_async(context, buf)
+
+        val = buf.getvalue()
+        if self.forced_output or not val.isspace():
+            buffer.write(val)
 
         return rendered